refactor: enum 패키지 이동
