+++ src/main/java/nextstep/interfaces/interceptor/LoginInterceptor.java
@@ -1,21 +1,19 @@
 package nextstep.interfaces.interceptor;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.domain.model.template.Role;
+import nextstep.domain.model.template.enumeration.Role;
 import nextstep.domain.model.template.annotation.NoAuth;
 import nextstep.infra.jwt.AuthorizationExtractor;
 import nextstep.infra.jwt.JwtTokenProvider;
 import nextstep.infra.exception.auth.AuthorizationException;
 import nextstep.infra.exception.auth.NoAccessAuthorityException;
 import nextstep.infra.exception.auth.NoSuchTokenException;
-import org.springframework.stereotype.Component;
 import org.springframework.util.ObjectUtils;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.util.logging.Handler;
 
 @RequiredArgsConstructor
 public class LoginInterceptor implements HandlerInterceptor {
